Updated Room Functionality
Made room and enemy prefabs which the game maneger now stores in arrays.
when you hit play the GM will load a random room and the room will populate it with random enemies(theres only one enemy lol)
there is a bug where it creates 1 extra enemy, not sure why.